+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-# 1.0.0 (2021-10-12)
+# 1.0.0 (2021-11-02)
 
 
 ### Features
 
 * **add update-changelog script:** Add script to package.json ([49ea3da](https://github.com/react-native-libraries/cz-jira-smart-validated/commit/49ea3da6228dec72959046ad29f71159c0cf7e66))
-* **engine.js:** add validations and set questions to use in this adapter ([7d587b1](https://github.com/react-native-libraries/cz-jira-smart-validated/commit/7d587b1cb524297050b992ba6dc29ede870c2fb0))
+* **engine.js:** add "nothing" option to workflow options ([3ee3564](https://github.com/react-native-libraries/cz-jira-smart-validated/commit/3ee3564708c9a2023d674bdff396b234231c8219))
+* **engine.js:** add validations and set questions to use in this adapter ([4a7b249](https://github.com/react-native-libraries/cz-jira-smart-validated/commit/4a7b2491fabe675b47ba1f1e23198e6ffbb8c2fc))
 * **index.js:** add new options ([62f764e](https://github.com/react-native-libraries/cz-jira-smart-validated/commit/62f764ef186cd7718bcec5c163f9c9407335979a))
 * **package.json .eslintrc.json .husky .eslintignore:** first Commit ([f400587](https://github.com/react-native-libraries/cz-jira-smart-validated/commit/f400587c2e81fa4043a4e80d7e8d14749a05497c))
 * **package.json:** Add flag to ignore hooks ([8c9519c](https://github.com/react-native-libraries/cz-jira-smart-validated/commit/8c9519cee85884ec356696cd6cf4e6b21a9a6f37))